fix(cli): move autoresponder config subcommand into autoresponder_commands module
The config subcommand was left in cli.py after extraction, causing a
NameError for _get_autoresponder_config_manager which now lives in the
commands module. Move the command definition to the correct file.
