+++ src/dopemux/cli.py
@@ -2781,101 +2781,6 @@ def health(
         sys.exit(1)
 
 
-@autoresponder.command("config")
-@click.option(
-    "--enabled/--disabled",
-    default=None,
-    help="Enable or disable auto responder",
-)
-@click.option(
-    "--terminal-scope",
-    type=click.Choice(["current", "all", "project"]),
-    help="Terminal monitoring scope",
-)
-@click.option("--delay", type=float, help="Response delay in seconds (0-10)")
-@click.option("--timeout", type=int, help="Auto-stop timeout in minutes")
-@click.option(
-    "--whitelist/--no-whitelist",
-    default=None,
-    help="Enable/disable tool whitelisting",
-)
-@click.option(
-    "--debug/--no-debug",
-    default=None,
-    help="Enable/disable debug mode",
-)
-@click.pass_context
-def autoresponder_config(
-    ctx, enabled, terminal_scope, delay, timeout, whitelist, debug
-):
-    """
-    ⚙️ Configure auto responder settings
-
-    Update configuration options for Claude Auto Responder integration.
-    """
-    project_path = Path.cwd()
-    config_manager = _get_autoresponder_config_manager(ctx, project_path)
-
-    updates = {}
-    if enabled is not None:
-        updates["enabled"] = enabled
-    if terminal_scope:
-        updates["terminal_scope"] = terminal_scope
-    if delay is not None:
-        updates["response_delay"] = delay
-    if timeout:
-        updates["timeout_minutes"] = timeout
-    if whitelist is not None:
-        updates["whitelist_tools"] = whitelist
-    if debug is not None:
-        updates["debug_mode"] = debug
-
-    if not updates:
-        # Show current config
-        current_config = config_manager.get_claude_autoresponder_config()
-
-        table = Table(title="🤖 Auto Responder Configuration")
-        table.add_column("Setting", style="cyan")
-        table.add_column("Value", style="green")
-
-        table.add_row("Enabled", "Yes" if current_config.enabled else "No")
-        table.add_row("Terminal Scope", current_config.terminal_scope)
-        table.add_row("Response Delay", f"{current_config.response_delay}s")
-        table.add_row("Timeout", f"{current_config.timeout_minutes} minutes")
-        table.add_row(
-            "Whitelist Tools", "Yes" if current_config.whitelist_tools else "No"
-        )
-        table.add_row("Debug Mode", "Yes" if current_config.debug_mode else "No")
-
-        console.logger.info(table)
-        return
-
-    # Apply updates
-    try:
-        config_manager.update_claude_autoresponder(**updates)
-        console.logger.info("[green]✅ Configuration updated[/green]")
-
-        # Show what changed
-        for key, value in updates.items():
-            console.logger.info(f"[blue]  {key}: {value}[/blue]")
-
-        # Restart if running
-        if (project_path / ".dopemux").exists():
-            from integrations.claude_autoresponder import create_autoresponder_manager
-
-            autoresponder_manager = create_autoresponder_manager(
-                config_manager, project_path
-            )
-            if autoresponder_manager.is_running():
-                console.print(
-                    "[yellow]🔄 Restarting auto responder with new settings...[/yellow]"
-                )
-                autoresponder_manager.restart()
-
-    except ValueError as e:
-        console.logger.error(f"[red]❌ Configuration error: {e}[/red]")
-        sys.exit(1)
-
 
 def _get_attention_emoji(state: Optional[str]) -> str:
     """Get emoji for attention state."""
